Fix missing SimplePlayer import
